+++ src/app.scss
@@ -8,6 +8,7 @@
   margin: 0 auto;
   width: 80%;
 }
+
 body {
   background-color: $primary-violet;
 }
